Revert "Merge pull request #49 from ISISComputingGroup/update_to_base_256babf_2021_6_15"
This reverts commit 24e071cca8705113f089638f91ff967c9aa9454d, reversing
changes made to f00612f4cec8a157fe32487204da0f0e56dbb0ae.

+++ modules/database/src/ioc/dbtemplate/msi.cpp
@@ -3,7 +3,6 @@
 * National Laboratory.
 * Copyright (c) 2002 The Regents of the University of California, as
 * Operator of Los Alamos National Laboratory.
-* SPDX-License-Identifier: EPICS
 * EPICS Base is distributed subject to a Software License Agreement found
 * in the file LICENSE that is included with this distribution.
 \*************************************************************************/
@@ -301,11 +300,11 @@ static void makeSubstitutions(inputData * const inputPvt,
             char *pstart;
             char *pend;
             int  cmdind=-1;
-            size_t  i;
+            int  i;
 
             for (i = 0; i < NELEMENTS(cmdNames); i++) {
                 if (strstr(command, cmdNames[i])) {
-                    cmdind = (int)i;
+                    cmdind = i;
                 }
             }
             if (cmdind < 0) goto endcmd;
